cast to quiet msvc compiler.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@9445 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/scapi.c
@@ -276,7 +276,7 @@ sc_generate_keyed_hash(const oid * authtype, size_t authtypelen,
     if (buf_len != properlength) {
         QUITFUN(rval, sc_generate_keyed_hash_quit);
     }
-    if (*maclen > buf_len)
+    if ((int)*maclen > buf_len)
         *maclen = buf_len;
     memcpy(MAC, buf, *maclen);
 #else
